Fix remaining CI build errors for all platforms
Windows - H.264 Profile Constants:
- FF_PROFILE_H264_* not compile-time constants in older FFmpeg
- Changed from switch to if-else with #ifdef guards
- Same issue as HEVC profiles

Linux - Qt Type Conversions (Qt 6.5.3):
- int64_t to QVariant ambiguous in explorerpanel.cpp
- Cast to qlonglong explicitly: static_cast<qlonglong>(atom.offset)

Linux - Qt Signal Compatibility (Qt 6.5.3):
- QCheckBox::checkStateChanged doesn't exist (added in Qt 6.7)
- Use QCheckBox::stateChanged instead (available in Qt 6.5.3)
- Both signals work the same way for checkbox state changes

All fixes maintain compatibility:
- FFmpeg 4.x/5.x/6.x/7.x (profile constants)
- Qt 6.5.3 (CI) and Qt 6.11+ (local dev)

+++ src/panels/explorerpanel.cpp
@@ -253,8 +253,8 @@ void ExplorerPanel::addMP4AtomNodes(QTreeWidgetItem* parent, const QVector<MP4At
         atomItem->setText(2, QString::number(atom.percentage, 'f', 2));
         atomItem->setCheckState(0, Qt::Checked);
 
-        // Store atom offset as user data
-        atomItem->setData(0, Qt::UserRole, atom.offset);
+        // Store atom offset as user data (cast to qlonglong for Qt 6.5.3)
+        atomItem->setData(0, Qt::UserRole, static_cast<qlonglong>(atom.offset));
 
         // Recursively add children
         if (!atom.children.isEmpty()) {
